+++ dotnet-advisor-core/Payloads/AdvisorCoreConfig.cs
@@ -2,11 +2,11 @@ using Microsoft.Extensions.Logging;
 
 namespace StormGeo.AdvisorCore.Payloads;
 
-public class AdvisorCoreConfig(string token, int attempts, int delay, Dictionary<string, string> headers, ILogger? logger)
+public class AdvisorCoreConfig(string token, int attempts, double delay, Dictionary<string, string> headers, ILogger? logger)
 {
     public string Token { get; } = token;
     public int Attempts { get; } = attempts;
-    public int Delay { get; } = delay;
+    public double Delay { get; } = delay;
     public Dictionary<string, string> Headers { get; } = headers;
     public ILogger? Logger { get; } = logger;
 }
